Rename edsl/data module to edsl/caching
This commit renames the data module to better reflect its purpose as a caching module:
- Moved all files using git mv to preserve history
- Updated all import statements throughout the codebase from edsl.data to edsl.caching
- Updated doctests and examples with new import paths
- Updated documentation URLs
- All tests are passing, but some integration tests may still require additional modifications
