+++ edsl/results/results.py
@@ -217,7 +217,7 @@ class Results(UserList, ResultsOperationsMixin, Base):
         self.completed = True
         self._fetching = False
         super().__init__(data)
-        from ..data import Cache
+        from ..caching import Cache
         from ..tasks import TaskHistory
 
         self.survey = survey
@@ -528,7 +528,7 @@ class Results(UserList, ResultsOperationsMixin, Base):
         include_task_history: bool = False,
         include_cache_info: bool = True,
     ) -> dict[str, Any]:
-        from ..data import Cache
+        from ..caching import Cache
 
         if sort:
             data = sorted([result for result in self.data], key=lambda x: hash(x))
@@ -640,7 +640,7 @@ class Results(UserList, ResultsOperationsMixin, Base):
         True
         """
         from ..surveys import Survey
-        from ..data import Cache
+        from ..caching import Cache
         from ..results import Result
         from ..tasks import TaskHistory
         from ..agents import Agent
@@ -1394,7 +1394,7 @@ class Results(UserList, ResultsOperationsMixin, Base):
         :param debug: if False, uses actual API calls
         """
         from ..jobs import Jobs
-        from ..data import Cache
+        from ..caching import Cache
 
         c = Cache()
         job = Jobs.example(randomize=randomize)
