+++ edsl/study/Study.py
@@ -8,7 +8,7 @@ from datetime import datetime
 from typing import Dict, Optional, Union
 from uuid import UUID, uuid4
 
-from ..data import Cache
+from ..caching import Cache
 #from edsl import set_session_cache, unset_session_cache
 from ..utilities.utilities import dict_hash
 from ..study.ObjectEntry import ObjectEntry
